select debugger gem dependent on ruby version - 2nd try
pry-debugger depends on debugger, which cannot be installed on
ruby>=2.0. An extensive, lengthy and sometimes unerving discussion can
be found on https://github.com/cldwalker/debugger/issues/125

On May 16, 2014, cldwalker suggested moving to byebug for
rubies>=2. This is reflected in the change to Gemfile.

As it turns out, travisci has problems working with the mri platform
directive, so bacd7829ee2ce86c3242c7aa45c987cdb3fafcde has been
reworked to resort to simple if-else logic.